+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/CHANGELOG.md
@@ -148,6 +148,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Breaking changes
 
+- Renamed module "Schema for the Admin" to "Schema Admin Fields"
 - Renamed all the "admin" fields: instead of prepending them with "unrestricted", now they are appended "ForAdmin"
 - The Access Control and Cache Control configuration lists will be broken: all fields for all non-root types broken will appear under "(Undefined entries)". These lists must be recreated